refactor(test): moved setup class and deleted iframetoken test
