+++ YulsnApiClient.Test/ContactsTest.cs
@@ -6,6 +6,7 @@ using System.Threading.Tasks;
 using Xunit;
 using YulsnApiClient.Client;
 using YulsnApiClient.Models.V1;
+using YulsnApiClient.Test.Abstractions;
 
 namespace YulsnApiClient.Test
 {
