+++ YulsnApiClient.Test/EmailDispatchesTest.cs
@@ -3,6 +3,7 @@ using System.Threading.Tasks;
 using Xunit;
 using YulsnApiClient.Client;
 using YulsnApiClient.Models.V1;
+using YulsnApiClient.Test.Abstractions;
 
 namespace YulsnApiClient.Test
 {
